requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
requests==2.31.0
python-dotenv==1.0.1

langchain==0.2.14
langgraph==0.2.1
langchain-openai==0.1.22
langsmith==0.1.104

openai==1.40.0
httpx==0.27.0
sseclient-py==1.8.0